requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.32.0
argon2-cffi==25.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
jinja2==3.1.4
aiofiles==24.1.0
psutil==6.1.0
redis==5.0.1
gunicorn==21.2.0